+++ src/modules/modem/analog/ModemAM.cpp
@@ -30,7 +30,7 @@ void ModemAM::demodulate(ModemKit *kit, ModemIQData *input, AudioThreadInput *au
         return;
     }
     
-	for (int i = 0; i < bufSize; i++) {
+	for (size_t i = 0; i < bufSize; i++) {
 		ampmodem_demodulate(demodAM, input->data[i], &demodOutputData[i]);
 	}
